Remove products from cart

+++ app/controllers/purchases_controller.rb
@@ -8,13 +8,22 @@ class PurchasesController < ApplicationController
     @user = User.find(params[:user_id])
     @purchase = Purchase.new(user: @user)
     @purchase.save
+    #if not @user.carts.empty?
+    purchased = false
     @user.carts.each do |c|
       pp = PurchaseProduct.new
-      pp.amount = c.amount
-      pp.product_id = c.product_id
-      pp.purchase = @purchase
-      pp.save
-      c.destroy
+      product = Product.find(c.product_id)
+      if product.update(stock: product.stock - c.amount)
+        purchased = true
+        pp.amount = c.amount
+        pp.product_id = c.product_id
+        pp.purchase = @purchase
+        pp.save
+        c.destroy
+      end
+    end
+    if not purchased
+      @purchase.destroy
     end
     redirect_to @user
   end
